cash/requirements.txt
requests
pandas
numpy
ta
PyYAML
colorlog